Add files via upload
This HTML code defines a basic webpage two share buttons for LinkedIn and WhatsApp. The functionality of the JavaScript code is to enable sharing the current webpage URL and title on LinkedIn and WhatsApp respectively when the corresponding buttons are clicked. It achieves this by opening a new window with the share URL constructed using the encodeURIComponent function to encode the URL parameters.